requirements.txt
openai
python-nmap
python-whois
typer
flask
pymetasploit3